+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Added `#delete!` method to `Source` model to allow deletion of sources
 - Added `#find`, `#find!`, `#find_by_md5_hash`, and `#find_by_md5_hash!` methods to
   `SourcesCollection` to find a source by one of its attributes
+- Added `#content` method to `Source` model to retrieve a source its content.
 
 ## Version: 0.3.0 (alpha)